+++ src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import './App.css';
-import Header from './Header'; 
+import Header from './Header/Header.jsx'; 
+import './Header/Header.css';
 
 function App() {
   return (
