+++ src/Header/Header.css
@@ -1,0 +1,12 @@
+.Header-app-bar {
+    background-color: black; 
+}
+
+.header-title {
+    font-size: 28px; 
+    font-weight: bold; 
+}
+
+.header-button:hover {
+    color: gray;
+}
